p29.29 refactor: thread ids.TransactionID everywhere
Convert the transaction entity id from bare int64 to the distinct ids.TransactionID
type across every layer: sqlc models (transactions.ID, splits.TransactionID FK,
SplitsVersion.TransactionID, TransactionsVersion.EntityID) + the txn/split queries
(Get/SoftDelete/InsertTransaction return/version params/UpdateTransactionParams.ID/
SplitVersionHistory/SplitsByTransaction/TransactionVersionAsOf/reconciliation rows/
balances+funds ledger rows); the store seam (Post/Update/Delete/GetTransaction,
TransactionHistory/AsOf/Splits, postTransactionTx + PostImportRow +
PostAndConvertExpenseReport + ConvertExpenseReport returns/params, the store-facing
Register/FundLedger/Drill + reconciliation TxnID fields, import-row PostedTxnID via
ids.Ptr); the reports layer (TransactionID alias + Cell.TxnID/WithTxn + the
revenueTxn map); synth/fixture builders (EditedTxn/DeletedTxn/MayRentTxn/
JuneDonationTxn, post/editTxn, UnclearedTxns); the ledgerimport CLI (tidTxns); the
web layer (editor/history/void/duplicate/register/drill/expense-convert parse
boundaries + txn-keyed counter maps + render-only TxnID fields); and the full test
fan-out.

Purely types, zero behavior change: goldens byte-identical, make check clean.
Boundary discipline (rule 15) — bare int64 only at the SQL scan + HTTP parse seams
(ids.TransactionID(parseID(s))); nullable posted_transaction_id stays
sql.NullInt64 in the sqlc model and types at the store boundary (ids.Ptr/ids.Null);
render-only web fields + itoa/FormatInt URL builds + driver.Value binds keep int64
and cast with int64(x); split ids (splits.ID) + account ids stay int64 (later
commits). ST1019: the balances/transactions/fixture test files that shadow `ids`
with a local var use the existing entids alias or a renamed local (sids/txnIDs).

+++ internal/store/imports_test.go
@@ -69,7 +69,7 @@ func TestPostImportRowCreatesBalancedTxnAndLinks(t *testing.T) {
 	}
 
 	// The txn is a real versioned ledger entry.
-	testutil.AssertVersioned(t, s.db, "transactions", txnID, "create")
+	testutil.AssertVersioned(t, s.db, "transactions", int64(txnID), "create")
 
 	// The row is LINKED: status=posted, posted_transaction_id = txnID.
 	row, err := s.GetImportRow(context.Background(), rowID)
